Rename core.py -> context.py
