+++ strax/__init__.py
@@ -17,7 +17,7 @@ from .config import *
 from .plugin import *
 from .mailbox import *
 from .processor import *
-from .core import *
+from .context import *
 
 from .io import *
