+++ strax/context.py
